Re-implementing Google Health using the ASTM CCR Standard, as it turns out the CCR Subset (CCRg) is in a direction that will continue to be complementary to the ASTM CCR Standard. Existing customers use the ASTM CCR Standard for importing data into Google Health. Although only items specified in the CCRg are supported in terms of retaining data as well as its display in the UI.
Using the ASTM CCR Standard does break the Google Health ccr model. In fact, it is recommended to follow the ASTM CCR XSD when leveraging Google Health. This is the first stab at bringing that closer to a reality.

Also included in this update is the ability to call methods directly from either the H9Service or the HealthService library. Using HealthService will authenticate against the Health Feeds service and use H9Service to call the H9 Feeds. Although, there is nothing stopping using the base query and providing the appropriate feeds, these default services allow for the new methods such as GetProfiles, GetCareRecords, GetNotices and PostNotice to map to the appropriate base address given the health service type.